Reworked audio system to be more inline with the way other systems work. Also various bugfixes included.

+++ kohi/code/kohi.core/src/core_render_types.h
@@ -453,7 +453,7 @@ typedef struct kstatic_mesh_submesh_render_data {
  */
 typedef struct kstatic_mesh_render_data {
     /** The identifier of the mesh instance being rendered. */
-    u64 instance_id;
+    u16 instance_id;
     /** @brief The number of submeshes to be rendered. */
     u32 submesh_count;
     /** @brief The array of submeshes to be rendered. */
